+++ src/pages/auth/loginPage.tsx
@@ -1,18 +1,12 @@
-import { useLoginMutation } from '@/api/auth/login/login.hooks';
-import { LoginResponseDto } from '@/api/auth/login/loginType';
 import googleLogo from '@/assets/auth/google_login.svg';
 import kakaoLogo from '@/assets/auth/kakao_logo.png';
 import naverLogo from '@/assets/auth/naver_logo.svg';
 import Button from '@/components/common/Button';
 import Input from '@/components/common/Input';
-import { useToast } from '@/hooks/useToast';
-import { ApiErrorResponseDto } from '@/types/apiErrorType';
-import { zodResolver } from '@hookform/resolvers/zod';
-import axios from 'axios';
-import { Cookies } from 'react-cookie';
-import { FormProvider, useForm } from 'react-hook-form';
-import { Link, useNavigate } from 'react-router-dom';
-import { z as zod } from 'zod';
+import { useLogin } from '@/hooks/login/useLogin';
+import { FormProvider } from 'react-hook-form';
+import { Link } from 'react-router-dom';
+
 //소셜로그인 버튼
 const socialLogin = [
   { name: 'google', src: googleLogo, url: '/' },
@@ -20,70 +14,13 @@ const socialLogin = [
   { name: 'naver', src: naverLogo, url: '/' },
 ];
 
-// 로그인 폼 스키마 정의
-const loginFormSchema = zod.object({
-  // 이메일 형식 지정
-  email: zod.string().email({ message: '이메일 형식이 아닙니다.' }),
-  password: zod.string().min(1, { message: '비밀번호를 입력해주세요.' }),
-});
-
 const LoginPage = () => {
-  const navigate = useNavigate();
-  const { showToast } = useToast();
-
-  const cookies = new Cookies();
-
-  const form = useForm({
-    resolver: zodResolver(loginFormSchema),
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-    mode: 'onSubmit',
-  });
+  const { form, handleLogin } = useLogin();
 
   const {
     register,
     formState: { errors },
-    getValues,
   } = form;
-
-  /**로그인 함수 */
-  const { mutate: LoginMutation } = useLoginMutation({
-    onSuccess: (data: LoginResponseDto) => {
-      if (data.user.role === 'student') {
-        navigate('/student-main', { replace: true });
-      } else if (data.user.role === 'teacher') {
-        navigate('/teacher-main', { replace: true });
-      }
-      cookies.set('accessToken', data.user.access_token);
-      cookies.set('refreshToken', data.user.refresh_token);
-    },
-    onError: (error) => {
-      console.error('Login Error:', error); // 에러 상세 로깅
-
-      // Axios 에러인 경우 더 상세한 로깅
-      if (axios.isAxiosError(error)) {
-        console.error('Axios Error Details:', {
-          response: error.response?.data,
-          status: error.response?.status,
-          headers: error.response?.headers,
-        });
-      }
-
-      const apiError = error as ApiErrorResponseDto;
-      const errorMessage =
-        apiError?.response?.data?.message ||
-        '로그인에 실패하였습니다. 다시 시도해주세요.';
-      showToast(errorMessage);
-    },
-  });
-
-  const handleLogin = async () => {
-    console.log('Login Attempt:', getValues());
-    const form = getValues();
-    LoginMutation(form);
-  };
 
   return (
     <FormProvider {...form}>
@@ -133,7 +70,7 @@ const LoginPage = () => {
               <p className='cursor-pointer'>아이디 찾기</p>
             </Link>
             <span className='inline-block h-[14px] w-[1px] bg-inputBorderColor'></span>
-            <Link to={'/find/password'}>
+            <Link to={'/reset/password'}>
               <p className='cursor-pointer'>비밀번호 찾기</p>
             </Link>
             <span className='inline-block h-[14px] w-[1px] bg-inputBorderColor'></span>
